+++ src/libs/Detach/Collisions/Geometry3D.cs
@@ -1180,6 +1180,32 @@ public static class Geometry3D
 		return false;
 	}
 
+	public static bool SphereCastCylinder(SphereCast sphereCast, Cylinder cylinder)
+	{
+		Vector2 cylinderXz = new(cylinder.BottomCenter.X, cylinder.BottomCenter.Z);
+		Vector2 rayOriginXz = new(sphereCast.Start.X, sphereCast.Start.Z);
+		Vector2 rayDirectionXz = new(sphereCast.End.X - sphereCast.Start.X, sphereCast.End.Z - sphereCast.Start.Z);
+
+		Vector2 d = cylinderXz - rayOriginXz;
+		float a = Vector2.Dot(rayDirectionXz, rayDirectionXz);
+		float b = 2 * Vector2.Dot(rayDirectionXz, d);
+		float c = Vector2.Dot(d, d) - (cylinder.Radius + sphereCast.Radius) * (cylinder.Radius + sphereCast.Radius);
+
+		float discriminant = b * b - 4 * a * c;
+		if (discriminant < 0)
+			return false;
+
+		float t1 = (-b + MathF.Sqrt(discriminant)) / (2 * a);
+		float t2 = (-b - MathF.Sqrt(discriminant)) / (2 * a);
+
+		float y1 = sphereCast.Start.Y + (sphereCast.End.Y - sphereCast.Start.Y) * t1;
+		float y2 = sphereCast.Start.Y + (sphereCast.End.Y - sphereCast.Start.Y) * t2;
+
+		return
+			y1 >= cylinder.BottomCenter.Y && y1 <= cylinder.BottomCenter.Y + cylinder.Height ||
+			y2 >= cylinder.BottomCenter.Y && y2 <= cylinder.BottomCenter.Y + cylinder.Height;
+	}
+
 	#endregion Sphere casting
 
 	#region Public helpers